+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 - Updated error message when the `.load()` method is provided with a nonexistent model path or an invalid file format.
 - Updated the `IntentClassifier` base model.
+- Updated the output model's directory structure for better organization.
 
 ## Release v0.3.2 - October 2, 2025